% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_add.R
\name{geom_add}
\alias{geom_add}
\title{geom_circ}
\usage{
geom_add(mapping = NULL, data, geom, offset = 0, ...)
}
\arguments{
\item{mapping}{aes mapping for 'geom'}

\item{data}{data to plot by 'geom', the column contained tree tip labels
should be as y in mapping.}

\item{geom}{geom function to plot the data.}

\item{offset}{numeric, distance between panels.}

\item{...}{additional parameters for 'geom'}
}
\value{
ggplot objec
}
\description{
plot tree associated data in another method.
}
\details{
'geom_circ()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping' and other parameters,
and align the graph with the tree 'p' side by side. '
}
\examples{
library(ggtree)
library(ggplot2)
tr <- rtree(10)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(10)))
p <- ggtree(tr)
p <- p + geom_add(data=dd, 
                   geom=geom_bar, 
                   mapping=aes(x=value, y=id),
                   orientation="y",
                   stat="identity",
                   position=position_stackx())
}
\author{
Shuangbin Xu and Guangchuang Yu
}
