% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axis_text.R
\name{geom_axis_text}
\alias{geom_axis_text}
\title{geom_axis_text}
\usage{
geom_axis_text(angle = 0, size = 0.8, nlayer = NULL, nbreak = 4, ...)
}
\arguments{
\item{angle}{numeric, the angle of text, default is 0 .}

\item{size}{numeric, the size of text, default is 0.8 .}

\item{nlayer}{integer, the specified layer of geom_fruit, default is NULL,
which mean the last layer.}

\item{nbreak}{integer, when the axis x is continuous, default is 4.}

\item{...}{additional parameters for 'geom_text'.}
}
\description{
add the axis of layer of geom_fruit.
}
\details{
'geom_axis_text' was designed to add the axis of layer of geom_fruit using
'geom_text'. So the dot parameters can refer the arguments of 'geom_text',
except \code{data} and \code{mapping}.
}
\examples{
library(ggplot2)
library(ggtree)
library(ggstar)
set.seed(1024)
tr <- rtree(100)
dd <- data.frame(id=tr$tip.label, value=abs(rnorm(100)))
dt <- data.frame(id=tr$tip.label, group=c(rep("A", 50), rep("B", 50)))
p <- ggtree(tr, layout="fan", open.angle=10)
p <- p \%<+\% dd \%<+\% dt
p1 <- p + geom_fruit(geom=geom_star,
                     mapping=aes(y=id, fill=group),
                     size=2.5,
                     starstroke=0) +
          geom_fruit(geom=geom_bar,
                     mapping=aes(x=value, y=id),
                     orientation="y",
                     stat="identity") +
          geom_axis_text(angle=-45, nudge_x=0.1)
}
\author{
Shuangbin Xu
}
