% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fruit.R
\name{geom_fruit_list}
\alias{geom_fruit_list}
\title{geom_fruit_list}
\usage{
geom_fruit_list(fruit, ...)
}
\arguments{
\item{fruit}{the layer of geom_fruit.}

\item{..., }{another layers of geom_fruit.}
}
\value{
ggplot object
}
\description{
add the layers to the same position out of ggtree.
}
\author{
Shuangbin Xu and GuangChuang Yu
}
