% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_totree.R
\name{geom_totree}
\alias{geom_totree}
\title{geom_totree}
\usage{
geom_totree(
  mapping = NULL,
  data,
  geom,
  offset = 0.03,
  pwidth = 0.2,
  addbrink = FALSE,
  linesize = 0.1,
  linecol = "grey50",
  ...
)
}
\arguments{
\item{mapping}{aes mapping for 'geom'}

\item{data}{data to plot by 'geom', the column contained tree tip labels
should be as y in mapping.}

\item{geom}{geom function to plot the data.}

\item{offset}{numeric, distance between panels, the ratio of distance
to tree, default is 0.03.}

\item{pwidth}{numeric, the width of new geom layer, this is the ratio of the
new geom to tree, default is 0.2.}

\item{addbrink}{logical, whether add the brink of different layers,
default is FALSE.}

\item{linesize}{numeric, the size of line of brink when 'addbrink'
is TRUE, default is 0.1.}

\item{linecol}{character, the color of line of brink when 'addbrink'
is TRUE, default is "grey50".}

\item{...}{additional parameters for 'geom'}
}
\value{
ggplot object
}
\description{
plot tree associated data in another method.
}
\details{
'geom_totree()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping'
and other parameters, and align the graph with the tree 'p' side by side.
}
\examples{
library(ggtree)
library(ggplot2)
library(ggstar)
tr <- rtree(100)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(100)))
dt = data.frame(id=tr$tip.label, group=c(rep("A",50),rep("B",50)))
p <- ggtree(tr, layout="circular")
p <- p + geom_totree(data=dt,
                  geom=geom_star,
                  mapping=aes(y=id, fill=group),
                  size=2.5,
                  starstroke=0)
p <- p + geom_totree(data=dd, 
                  geom=geom_bar, 
                  mapping=aes(x=value, y=id),
                  orientation="y",
                  stat="identity",
                  position=position_stackx())  
}
\author{
Shuangbin Xu and Guangchuang Yu
}
