% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_dodgex.r, R/position_dodgex2.r,
%   R/position_identityx.r, R/position_points_jitterx.R,
%   R/position_points_sinax.R, R/position_stackx.R
\name{PositionDodgex}
\alias{PositionDodgex}
\alias{PositionDodgex2}
\alias{PositionIdentityx}
\alias{PositionPointsJitterx}
\alias{PositionRaincloudx}
\alias{PositionPointsSinax}
\alias{PositionStackx}
\title{PositionDodgex}
\description{
PositionDodgex

PositionDodgex2

PositionIdentityx

PositionPointsJitterx

PositionRaincloudx

PositionPointsSinax

PositionStackx
}
\author{
Shuangbin Xu
}
