% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{ggplot_add}
\alias{as.treedata}
\alias{td_unnest}
\alias{td_filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:ggplot_add]{ggplot_add()}}}

  \item{ggtree}{\code{\link[ggtree:td_filter]{td_filter}}, \code{\link[ggtree:td_unnest]{td_unnest}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{tidytree}{\code{\link[tidytree:as.treedata]{as.treedata()}}}
}}

