% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_add.R
\name{geom_add}
\alias{geom_add}
\title{geom_add}
\usage{
geom_add(
  mapping = NULL,
  data,
  geom,
  offset = 0.03,
  pratio = 0.2,
  addbrink = FALSE,
  linesize = 0.1,
  linecol = "grey50",
  tippoint = FALSE,
  ...
)
}
\arguments{
\item{mapping}{aes mapping for 'geom'}

\item{data}{data to plot by 'geom', the column contained tree tip labels
should be as y in mapping.}

\item{geom}{geom function to plot the data.}

\item{offset}{numeric, distance between panels, the ratio of distance
to tree, default is 0.03.}

\item{pratio}{numeric, the ratio of new geom to tree, default is 0.2.}

\item{addbrink}{logical, whether add the brink of different layers,
default is FALSE.}

\item{linesize}{numeric, the size of line of brink when 'addbrink'
is TRUE, default is 0.1.}

\item{linecol}{character, the color of line of brink when 'addbrink'
is TRUE, default is "grey50".}

\item{tippoint}{logical, whether add the tip point layer,
only for geom_point, geom_star or geom_symbol layer, default is FALSE.}

\item{...}{additional parameters for 'geom'}
}
\value{
ggplot object
}
\description{
plot tree associated data in another method.
}
\details{
'geom_add()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping'
and other parameters, and align the graph with the tree 'p' side by side.
}
\examples{
library(ggtree)
library(ggplot2)
tr <- rtree(10)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(10)))
p <- ggtree(tr)
p <- p + geom_add(data=dd, 
                  geom=geom_bar, 
                  mapping=aes(x=value, y=id),
                  orientation="y",
                  stat="identity",
                  position=position_stackx())
}
\author{
Shuangbin Xu and Guangchuang Yu
}
