% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fruit.R
\name{geom_fruit}
\alias{geom_fruit}
\alias{fruit_plot}
\title{geom_fruit}
\usage{
geom_fruit(
  mapping,
  data = NULL,
  geom,
  offset = 0.03,
  pwidth = 0.2,
  addbrink = FALSE,
  linesize = 0.1,
  linecol = "grey50",
  position = "auto",
  ...
)

fruit_plot(
  p,
  data = NULL,
  geom,
  mapping,
  offset = 0.03,
  pwidth = 0.2,
  addbrink = FALSE,
  linesize = 0.1,
  linecol = "grey50",
  position = "auto",
  ...
)
}
\arguments{
\item{mapping}{aes mapping for 'geom'}

\item{data}{data to plot by 'geom', the column contained tree tip labels
should be as y in mapping.}

\item{geom}{geom function to plot the data.}

\item{offset}{numeric, distance between panels, the ratio of distance
to tree, default is 0.03.}

\item{pwidth}{numeric, the width of new geom layer, this is the ratio of the
new geom to tree, default is 0.2.}

\item{addbrink}{logical, whether add the brink of different layers,
default is FALSE.}

\item{linesize}{numeric, the size of line of brink when 'addbrink'
is TRUE, default is 0.1.}

\item{linecol}{character, the color of line of brink when 'addbrink'
is TRUE, default is "grey50".}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function, default is 'auto'.}

\item{...}{additional parameters for 'geom'}

\item{p}{tree view}
}
\value{
ggplot object
}
\description{
plot tree associated data in another method.
}
\details{
'geom_fruit()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping'
and other parameters, and align the graph with the tree 'p' side by side.
The default position parameters is 'auto'. If you want to set manually. You can use
'position_stackx()' or 'position_dodgex()' for 'geom_bar', 'position_identityx()' for
'geom_tile', 'geom_point', 'geom_star', 'geom_symbol' or other layers using 'identity'
position in 'ggplot2', and 'position_dodgex()' or 'position_dodgex2()' for 'geom_boxplot'
'geom_violin' or other layers using 'dodge' position in 'ggplot2'.
}
\examples{
library(ggtree)
library(ggplot2)
library(ggstar)
set.seed(1024)
tr <- rtree(100)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(100)))
dt = data.frame(id=tr$tip.label, group=c(rep("A",50),rep("B",50)))
p <- ggtree(tr, layout="circular") 

p1 <- p + geom_fruit(data=dt,
                  geom=geom_star,
                  mapping=aes(y=id, fill=group),
                  size=2.5,
                  starstroke=0)
p2 <- p1 + geom_fruit(data=dd, 
                  geom=geom_bar, 
                  mapping=aes(x=value, y=id),
                  orientation="y",
                  stat="identity")  

p3 <- fruit_plot(p=p, 
                 data=dt, 
                 geom=geom_star,
                 mapping=aes(y=id, fill=group),
                 size=2.5, starstroke=0)

p4 <- fruit_plot(p=p3,
                 data=dd,
                 geom=geom_bar,
                 mapping=aes(x=value, y=id),
                 orientation="y",
                 stat="identity")
p <- p \%<+\% dd \%<+\% dt
p5 <- p + geom_fruit(geom = geom_star,
                     mapping = aes(y=id, fill=group),
                     size = 2.5,
                     starstroke = 0) +
          geom_fruit(geom = geom_bar,
                     mapping = aes(x=value, y=id),
                     orientation = "y",
                     stat = 'identity')
}
\author{
Shuangbin Xu and Guangchuang Yu
}
