% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fruit.R
\name{geom_fruit}
\alias{geom_fruit}
\alias{fruit_plot}
\title{geom_fruit}
\usage{
geom_fruit(
  mapping,
  data = NULL,
  geom,
  offset = 0.03,
  pwidth = 0.2,
  position = "auto",
  grid.params = NULL,
  axis.params = list(axis = "none", text.angle = 0, text.size = 0.8, text = NULL, title
    = NULL, title.size = 3, title.height = 0.1, title.angle = 0, title.color = "black",
    nbreak = 4, line.size = 0.2, line.color = "grey", line.alpha = 1, ...),
  ...
)

fruit_plot(
  p,
  data = NULL,
  geom,
  mapping,
  offset = 0.03,
  pwidth = 0.2,
  position = "auto",
  ...
)
}
\arguments{
\item{mapping}{aes mapping for 'geom'}

\item{data}{data to plot by 'geom', the column contained tree tip labels
should be as y in mapping.}

\item{geom}{geom function to plot the data.}

\item{offset}{numeric, distance between panels, the ratio of distance
to tree, default is 0.03.}

\item{pwidth}{numeric, the width of new geom layer, this is the ratio of the
new geom to tree, default is 0.2.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function, default is 'auto'.}

\item{grid.params}{list, the parameters to control the attributes of grid lines,
default is NULL.}

\item{axis.params}{list, the parameters to control the attributes of pseudo axis.}

\item{...}{additional parameters for 'geom'

grid.params control the attributes of grid line, it can be referred to the
following parameters:
\itemize{
\item \code{vline} logical, whether add the vertical line, default is FALSE.
\item \code{color} color of line, default is grey.
\item \code{size} the width of line, default is 0.2.
\item \code{alpha} the colour transparency of line, default is 1.
\item \code{lineend} Line end style (round, butt, square), default is "butt".
\item \code{linejoin} Line end style (round, butt, square), default is "round".
\item \code{linetype} Type of line, default is 1.
}

axis.params control the attributes of pseudo axis, it can be referred to the
following parameters:
\itemize{
\item \code{axis} character, add the pseudo axis, "none" don't display axis (default),
"x" display the x axis, "y" display the (y) axis of tree tip, "xy" display the two axis.
\item \code{text} vector, the text of axis x, default is NULL, it is only valid when
the text of axis is single and x is discrete.
\item \code{vjust} numeric, A numeric specifying vertical justification, default is 0.5.
\item \code{hjust} numeric, A numeric specifying horizontal justification, default is 0.5.
\item \code{text.angle} numeric, the angle of axis text, default is 0.
\item \code{text.size} numeric, the size of axis text, default is 0.8.
\item \code{title} character, the title of panel or x-axis label, default is NULL, it is only
valid when "x" axis exists.
\item \code{title.size} numeric, the size of title text, default is 3.
\item \code{title.height} numeric, the height of title text position more than tree, default is 0.1,
it is relative to height of tree.
\item \code{title.angle} numeric, the angle of title text, default is 0.
\item \code{title.color} character, the color of title text, default is "black".
\item \code{nbreak} numeric, meaning the number of axis to break,
integer giving the \emph{desired} number of intervals. Non-integer values are rounded down.
It is only valid when x is continuous, default is 4.
\item \code{line.size} numeric, the size of axis line, default is 0.2.
\item \code{line.color} character, the color of axis color, default is "grey".
\item \code{line.alpha} numeric, the colour transparency of line, default is 1.
}}

\item{p}{tree view}
}
\value{
ggplot object
}
\description{
plot tree associated data in another method.
}
\details{
'geom_fruit()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping'
and other parameters, and align the graph with the tree 'p' side by side.
The default position parameters is 'auto'. If you want to set manually. You can use
'position_stackx()' or 'position_dodgex()' for 'geom_bar', 'position_identityx()' for
'geom_tile', 'geom_point', 'geom_star', 'geom_symbol' or other layers using 'identity'
position in 'ggplot2', and 'position_dodgex()' or 'position_dodgex2()' for 'geom_boxplot'
'geom_violin' or other layers using 'dodge' position in 'ggplot2'. The axis line and text
can be added using 'axis.params=list(axis="x",...)', and the grid line also can be added
using 'grid.params=list(...)'.
}
\examples{
library(ggtree)
library(ggplot2)
library(ggstar)
set.seed(1024)
tr <- rtree(100)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(100)))
dt = data.frame(id=tr$tip.label, group=c(rep("A",50),rep("B",50)))
p <- ggtree(tr, layout="circular") 

p1 <- p + 
      geom_fruit(
          data=dt,
          geom=geom_star,
          mapping=aes(y=id, fill=group),
          size=2.5,
          starstroke=0
      )
p2 <- p1 + 
      geom_fruit(
          data=dd, 
          geom=geom_bar, 
          mapping=aes(x=value, y=id),
          orientation="y",
          stat="identity"
      )  

p3 <- fruit_plot(
           p=p, 
           data=dt, 
           geom=geom_star,
           mapping=aes(y=id, fill=group),
           size=2.5, 
           starstroke=0
      )

p4 <- fruit_plot(
           p=p3,
           data=dd,
           geom=geom_bar,
           mapping=aes(x=value, y=id),
           orientation="y",
           stat="identity"
      )
p <- p \%<+\% dd \%<+\% dt
p5 <- p + 
      geom_fruit(
          geom = geom_star,
          mapping = aes(y=id, fill=group),
          size = 2.5,
          starstroke = 0
      ) +
      geom_fruit(
          geom = geom_bar,
          mapping = aes(x=value, y=id),
          orientation = "y",
          stat = 'identity'
      )
}
\author{
Shuangbin Xu and Guangchuang Yu
}
