% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_dodgex.r, R/position_dodgex2.r,
%   R/position_identityx.r, R/position_stackx.R
\docType{data}
\name{PositionDodgex}
\alias{PositionDodgex}
\alias{PositionDodgex2}
\alias{PositionIdentityx}
\alias{PositionStackx}
\title{PositionStackx}
\description{
PositionStackx
}
\author{
Shuangbin Xu
}
\keyword{datasets}
