% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_dodgex.r, R/position_dodgex2.r
\name{position_dodgex}
\alias{position_dodgex}
\alias{position_dodgex2}
\title{Dodge overlapping objects side-to-side}
\usage{
position_dodgex(
  width = NULL,
  hexpand = NA,
  vexpand = NA,
  preserve = c("total", "single")
)

position_dodgex2(
  width = NULL,
  preserve = c("total", "single"),
  hexpand = NA,
  vexpand = NA,
  padding = 0.1,
  reverse = FALSE
)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms.}

\item{hexpand}{numeric, Horizon expand for geoms that have a position, default is NA.}

\item{vexpand}{numeric, Vertical expand for geoms that have a position, default is NA.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{padding}{Padding between elements at the same position. Elements are
shrunk by this proportion to allow space between them. Defaults to 0.1.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\description{
Dodging preserves the vertical position of an geom while adjusting the
horizontal position. \code{position_dodge2} is a special case of \code{position_dodge}
for arranging box plots, which can have variable widths. \code{position_dodge2}
also works with bars and rectangles. But unlike \code{position_dodge},
\code{position_dodge2} works without a grouping variable in a layer.
}
\seealso{
Other position adjustments: 
\code{\link{position_identityx}()}
}
\concept{position adjustments}
