% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_identityx.r
\name{position_identityx}
\alias{position_identityx}
\title{adjust identity position.}
\usage{
position_identityx(hexpand = NA, vexpand = NA)
}
\arguments{
\item{hexpand}{numeric, Horizon expand for geoms that have a position, default is NA.}

\item{vexpand}{numeric, Vertical expand for geoms that have a position, default is NA.}
}
\description{
adjust identity position.
}
\seealso{
Other position adjustments: 
\code{\link{position_dodgex}()}
}
\concept{position adjustments}
