% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_points_sinax.R
\name{position_points_sinax}
\alias{position_points_sinax}
\title{adjust ridgeline plot position which can be shifted vertically or horizontally.}
\usage{
position_points_sinax(
  rel_min = 0.02,
  rel_max = 0.98,
  seed = NULL,
  hexpand = NA
)
}
\arguments{
\item{rel_min}{numeric, the relative minimum value at which a point can be placed.}

\item{rel_max}{numeric, the relative maximum value at which a point can be placed.}

\item{seed}{numeric, Random seed, if set to NULL, the current random number
generator is used. If set to NA, a new random random seed is generated.
If set to a number, this number is used as seed for jittering only,
default is NULL.}

\item{hexpand}{numeric, distance to be shifted horizontally for geoms that
have a position, default is NA.}
}
\value{
position method.
}
\description{
This is a position adjustment specifically for 'geom_density_ridges()', but it
add 'hexpand' that can control shift horizontally.
}
\seealso{
Other position adjustments: 
\code{\link{position_dodgex}()},
\code{\link{position_identityx}()}
}
\author{
Shuangbin Xu
}
\concept{position adjustments}
