% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_stackx.R
\name{position_stackx}
\alias{position_stackx}
\title{Stack overlapping objects (method2) on top of each another}
\usage{
position_stackx(vjust = 1, hexpand = NA, vexpand = NA, reverse = FALSE)
}
\arguments{
\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}

\item{hexpand}{numeric, Horizon expand for geoms that have a position, default is NA.}

\item{vexpand}{numeric, Vertical expand for geoms that have a position, default is NA.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\description{
Stack overlapping objects (method2) on top of each another
}
\author{
Shuangbin Xu
}
