% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot_add}
\alias{ggtree}
\alias{td_unnest}
\alias{td_filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{ggplot_add}}}

  \item{ggtree}{\code{\link[ggtree]{ggtree}}, \code{\link[ggtree]{td_filter}}, \code{\link[ggtree]{td_unnest}}}
}}

