% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{ggplot_add}
\alias{new_scale}
\alias{new_scale_fill}
\alias{new_scale_color}
\alias{new_scale_colour}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggnewscale}{\code{\link[ggnewscale]{new_scale}}, \code{\link[ggnewscale]{new_scale_color}}, \code{\link[ggnewscale]{new_scale_colour}}, \code{\link[ggnewscale]{new_scale_fill}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot_add}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

